+++ src/cernml/coi/checkers/_single_opt.py
@@ -10,7 +10,6 @@ import typing as t
 
 import gymnasium as gym
 import numpy as np
-from scipy.optimize import LinearConstraint, NonlinearConstraint
 
 from .._typeguards import is_single_optimizable
 from ..protocols import Constraint, SingleOptimizable
@@ -55,6 +54,8 @@ def assert_optimization_space(env: SingleOptimizable) -> None:
 
 def assert_constraints(constraints: t.Sequence[Constraint]) -> None:
     """Check that the list of constraints contains only constraints."""
+    from scipy.optimize import LinearConstraint, NonlinearConstraint
+
     allowed_types = (LinearConstraint, NonlinearConstraint)
     for constraint in constraints:
         assert isinstance(constraint, allowed_types), (